Fix iPXE boot and provisioning process
Not sure how this was working... and now is not but pxe is now failing to start
I am like 99.999999% sure I have had this running without http:// infront of the
URLs for over a month now but it doesn't work without it now. Added in some
more logic around provisioning and fixed a logic error because asset when none
exists returns false not nil.