Add "better" logging and default route for ipxe
This does some refactoring around how we log to collins just to
make it ever so slightly more clean... I also added a default route
for systems that don't have an asset tag set so it doesn't just look
like iPXE failed to load when it gets a 404. This should likely just
be made an optional param instead of part of the path.