Add an override for the default site config
This adds an easy way to specify a base config and then put a sparse
amount of info in other files on a per site basis. This also adds an
endpoint for making sure that your config is as you would expect.
This should not load the file every time the evie function is called
that is just silly but i'll fix it later.