Generate Vault Token By Nodeclass
This generates a vault token on request so that a server always gets
a fresh token. I need to learn a little bit more about vault and how
the token refresh works and how to handle this in the case that a token
isn't refreshed for X amount of time. We also need to put some auth in
place so someone can't just make requests against this endpoint to get
a valid vault token... But where to keep the password for that.